feat(import): restructure bookmark import flow and add edit capabilities
