+++ src/lib/stores/edit-bookmark.store.ts
@@ -1,4 +1,4 @@
-import type { Bookmark } from '$lib/types/Bookmark.type';
+import type { Bookmark, BookmarkEdit } from '$lib/types/Bookmark.type';
 import { writable } from 'svelte/store';
 
-export const editBookmarkStore = writable<Partial<Bookmark>>({});
+export const editBookmarkStore = writable<Partial<Bookmark> | BookmarkEdit>({});
