+++ src/lib/types/common/BulkList.type.ts
@@ -1,10 +1,3 @@
-import type { Metadata } from '../Metadata.type';
+import type { BookmarkEdit } from '../Bookmark.type';
 
-export type BulkListItem = Partial<Metadata> & {
-	id: number;
-	icon: string | null;
-	url: string;
-	title: string;
-	category: string;
-	selected: boolean;
-};
+export type BulkListItem = BookmarkEdit;
